Fix hipfb files not installed on the first build+install
A small oversight in the 109e7fe6 which made it so hipfb files are
not recompiled on every CMake run and that they detect changes in
dependencies.

The GLOB used for the install target runs at the configure time,
so it was missing these files as they are generated at the compile
time.

Use explicit list of hipfb files, which solves this issue, and
also avoids install of hipfb used for the regression tests.