Skip installation of bvh_build_array.h
Previous fix in this area made it so it consistently installed,
regardless of whether it is a fresh build install, or a subsequent
one. This was the safest as it was done to ensure Blender dependencies
are build in a reproducible manner, as we were reaching the release
phase.

Looking back into this topic: the header is quite big and is only
needed during compilation.